+++ src/drawer/components/WarningBox/WarningAccordion/WarningAccordion.style.ts
